hammerhead: disable or enable mpdecision dependend on the kernel
We introduce now an open source mpdecision variant. Thanks a ton to
showp1984 for it.
Problem is that it conflicts with googles prebuilt mpdecision even when
it is just disabled over the service. Only removing or renaming the correspondent
blob will let run our new msm_mpdecision as full replacement. As well this prevend
the user for doing weird other things.

Whatever we need to keep up 3rd party kernel compatibility. So we will check via
init.d script if the kernel supports msm_mpdecision...if yes we disable the prebuilt
if not we enable the prebuilt for the 3rd party kernel. All other the kernel and ramdisk
will handle later.

Change-Id: I27d9157607443635178ad169eebb3c625c55d669
Signed-off-by: Josue Rivera <prbassplayer@gmail.com>